feat(ui): improve `CoreFileInput`
The main issue of `CoreFileInput` was that, when it re-render, the files selected label is cleared and there is no way to set it again.

To prevent that, the easiest way is to hide the `<input type="file" />` and display instead a custom button and a label.

This unlock a cool possibility to create the link of the file and let the user download the file.