fix(ui): fix `CoreDataFrame` download for big file. WF-39
The issue was too many parameters was given to `String.fromCharCode()` and make and error like "Maximum call stack size exceeded".

In fact, there is no need to try to convert the CSV to base64. So I simply used `Blob` and then `URL.createObjectURL` instead which should be more efficient.