feat(ui): add filter to `BuilderSidebarToolbar`. WF-42
I introduced a component named `BuilderSidebarTitleSearch` which extract the existing search logic of `BuilderSidebarTree` and use it for `BuilderSidebarToolbar`.

I also took the opportunity to improve the accessibility since we use `<i>` elements as button, but they're not focusable and clickable with the keyboard.

And finally, I introduced an E2E test which cover the new feature.