fix(ui): display loader during async loading component
I rencently introduced Async Components but some of them are really big,
and the loading time is noticable. It renders as blank area which is not
good.

So I used `loadingComponent` property to fill the space during the
loading, and improved a bit the loading strategy to improve the UX.