chore(ui): improve hash parsing + test - WF-72
It's possible to use `URLSearchParams` instead of complex Regex. As
it'sa very testable part of the code, it was a good opportunity to
introduceVitest and cover it.This improvement made me realise that a bug
was present, and we didn'tserialize the params when converting back the
object to hash (`.key =length` was not called).