fix(ui): avoid reseting the "selected" state on all field change - WF-155
We want to disable the component "selected" state when the style of the
component is changed (like when the user changes the background). But
actually, this is done when **any** of `evaluatedFields` changes.

The fix was simply to run the behavior only when a "Style" field change.
So I just tweaked a bit the `watch` source in `ComponentProxy`.