fix(ui): avoid bundling `injectionKeys` in custom components
The custom component built using `npm run custom.build` was not working
correctly because the `provide/inject` failed.

The error was, we bundled `injectionKeys` in the `template.umd.js` , So
it makes defining two different `Symbol` which point to a different
`provide`.

It's because the `rollupOptions.external` in Vite config pointed to
`../injectionKeys`, but we sometime use the alias `@/injectionKeys`. So
I simply use `path.resolve` to be sure `injectionKeys` are not bundled
in custom components.