feat(ui): use floating UI for autocomplete - WF-233
Autocomplete needs to use `position: sticky` instead of `position:
absolute` because `BuilderTemplateInput` might be used inside a
container where the overflow is hidden.

Using floating UI also help the dropdown to be displayed on the top of
the input if there is no enough space on the bottom of the element.