chore(ui): bump library `vue-pdf` version
Bump `@tato30/vue-pdf` version to bump `pdf.js` version. It's needed to
patch a vulnerability that executes JavaScript encoded in the PDF file.

This behavior is blocked on writer.com that block this behavior with CSP, and makes the PDF not loading.

fixes https://github.com/writer/writer-framework/security/dependabot/22

> If pdf.js is used to load a malicious PDF, and PDF.js is configured with `isEvalSupported` set to `true` (which is the default value), unrestricted attacker-controlled JavaScript will be executed in the context of the hosting domain.