fix(writer): make `AppRunner.import_zip` compatible with file buffer - AB-251
Previously, we were removing the root folder when importing a project
from a zip in `AppRunner.import_zip`. It was an issue when the user run
the application with flags `--enable-file-buffering --buffer-path`. So I
introduced a new function that synchronizes the contents of the ZIP
folder to the APP path folder in a one-way manner.